+++ var/html/a.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+ <html lang="en">
+ <head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Upload File</title>
+ </head>
+ <body>
+  <form action="/cgi-bin/a" enctype="multipart/form-data" method="post">
+    <label for="filename">Upload file:</label>
+    <br>
+    <input type="file" name="filename">
+    <br>
+    <input type="submit" value="Upload">
+  </form>
+</body>
+ </html>
