+++ webserve/Server/Server.hpp
@@ -28,7 +28,6 @@
 // kqueue를 위한 헤더파일
 #include <sys/event.h>
 
-#include "../ServerController/ServerController.hpp"
 #include "../Connection/Connection.hpp"
 #include "../CommonLogFormat/CommonLogFormat.hpp"
 
@@ -38,21 +37,21 @@ class Server
 {
 private:
     pthread_t                   _s_thread;
+    ServerBlock*                _server_block;
+
     int                         _server_socket;
     struct sockaddr_in          _server_addr;
+
     std::vector<struct kevent>  _change_list;
     struct kevent*              _curr_event;
     std::map<int, Connection>   _clients;
     int                         _kqueue;
     struct kevent               _event_list[8];
-    ServerBlock*                _server_block;
     
     void socket_init(int port, unsigned int ip_addr);
     void change_events(uintptr_t ident, int16_t filter, uint16_t flags, uint32_t fflags, intptr_t data, void *udata);
     void disconnect_client(int client_fd, std::map<int, Connection> &clients);
     void accept_new_client();
-    void receiveMessage();
-    void sendMessage();
 
     static void* threadFunction(void *);
